# 0-allocation add_mul for Rational{BigInt}