```julia
using MutableArithmetics
using BenchmarkTools

function non_mutable()
    T = Rational{BigInt}
    a = one(T)
    b = one(T)
    c = one(T)
    @benchmark $a + $b * $c
end

function mutable()
    T = Rational{BigInt}
    a = one(T)
    b = one(T)
    c = one(T)
    buffer = buffer_for(add_mul, T, T, T)
    @benchmark buffered_operate!($buffer, add_mul, $a, $b, $c)
end

julia> non_mutable()
BenchmarkTools.Trial: 10000 samples with 550 evaluations.
 Range (min … max):  208.478 ns … 24.738 μs  ┊ GC (min … max):  0.00% … 59.11%
 Time  (median):     234.192 ns              ┊ GC (median):     0.00%
 Time  (mean ± σ):   399.755 ns ±  1.942 μs  ┊ GC (mean ± σ):  23.80% ±  4.82%

  ▄▅▄▅███▆▅▄▃▃▂▁▂▁▁▁                                           ▂
  ████████████████████▇▇▇▇▇▆▇███▇▆▅▅▅▅▄▅▅▆▅▆▆▅▃▃▄▃▃▃▃▅▃▃▁▃▅▄▅▃ █
  208 ns        Histogram: log(frequency) by time       474 ns <

 Memory estimate: 576 bytes, allocs estimate: 18.

# BEFORE THIS PR

julia> mutable()
BenchmarkTools.Trial: 10000 samples with 237 evaluations.
 Range (min … max):  315.295 ns … 257.897 μs  ┊ GC (min … max):  0.00% … 55.17%
 Time  (median):     348.675 ns               ┊ GC (median):     0.00%
 Time  (mean ± σ):   711.304 ns ±   9.011 μs  ┊ GC (mean ± σ):  26.71% ±  2.12%

           ▆█▃                                                   
  ▂▂▂▃▃▃▄▆▇████▅▄▄▃▃▃▃▃▂▂▂▂▂▂▂▂▃▃▃▃▃▃▃▃▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂ ▃
  315 ns           Histogram: frequency by time          500 ns <

 Memory estimate: 408 bytes, allocs estimate: 21.


# AFTER THIS PR

julia> mutable()
BenchmarkTools.Trial: 10000 samples with 927 evaluations.
 Range (min … max):  101.798 ns … 147.569 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     109.350 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   110.340 ns ±   4.182 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

             ▁▄▄▃▁▄▆█▅▄▁                                         
  ▂▂▂▃▂▂▁▃▂▄▇████████████▇▇▆▆▇▅▄▃▃▃▃▃▃▃▃▃▃▃▃▃▃▃▂▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂ ▄
  102 ns           Histogram: frequency by time          126 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

Closes https://github.com/jump-dev/MutableArithmetics.jl/issues/187